Create README.rst
I have added changes to promote inclusivity according to the policy and ethical standards of the project to help explain to novice users what a web crawler is and give them an example. I believe this is vital to the development community as it could promote the beginner to get started using the project.